Rename `internal` pkg to `awxtesting` to avoid confusing.
